backend/requirements.txt
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
bitarray==2.9.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
cryptography==43.0.0
cytoolz==0.12.3
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
parsimonious==0.10.0
pillow==10.4.0
protobuf==5.27.2
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pyunormalize==15.1.0
pywin32==306
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
rpds-py==0.19.1
sqlparse==0.5.1
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
web3==6.20.1
websockets==12.0
yarl==1.9.4
